# contracts/app-v1/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HeatmapGenerateRequest": {
        "additionalProperties": false,
        "properties": {
          "normalization_method": {
            "default": "SESSION_MAX",
            "enum": [
              "SESSION_MAX",
              "WINDOW_MAX"
            ],
            "title": "Normalization Method",
            "type": "string"
          },
          "window_from_t": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window From T"
          },
          "window_to_t": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window To T"
          }
        },
        "title": "HeatmapGenerateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Application API for CrowdSight recorded fixed-camera crowd monitoring. Provides media catalog, zone set versioning, session orchestration, synchronized playback queries, and relative image-space analytics. Fail-closed governance: operational alerts remain disabled in experimental views.",
    "title": "CrowdSight Application API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_api_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Session Api V1 Sessions  Session Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purge session, related records, and artifacts",
        "tags": [
          "Analysis Sessions"
        ]
      },
      "get": {
        "operationId": "get_session_api_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get session state and progress",
        "tags": [
          "Analysis Sessions"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/export": {
      "get": {
        "operationId": "export_session_api_v1_sessions__session_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "Export format",
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "jsonl",
              "description": "Export format",
              "enum": [
                "jsonl",
                "csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export session observation archive or CSV counts",
        "tags": [
          "Analytics & Export"
        ]
      }
    },
...
    "/api/v1/sessions/{session_id}/heatmaps": {
      "post": {
        "operationId": "generate_heatmap_api_v1_sessions__session_id__heatmaps_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeatmapGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeatmapMetadataModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate an image-space relative heat map",
        "tags": [
          "Analytics & Export"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/heatmaps/{artifact_id}": {
      "get": {
        "operationId": "get_heatmap_artifact_api_v1_sessions__session_id__heatmaps__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download heat map image or metadata",
        "tags": [
          "Analytics & Export"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/notes": {
      "get": {
        "operationId": "list_notes_api_v1_sessions__session_id__notes_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NoteResponseModel"
                  },
                  "title": "Response List Notes Api V1 Sessions  Session Id  Notes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List operator notes for session",
        "tags": [
          "Analytics & Export"
        ]
      },
      "post": {
        "operationId": "create_note_api_v1_sessions__session_id__notes_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NoteCreateModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteResponseModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create an operator note",
        "tags": [
          "Analytics & Export"
        ]
      }
    },
    "/api/v1/sessions/{session_id}/notes/{note_id}": {
      "delete": {
        "operationId": "delete_note_api_v1_sessions__session_id__notes__note_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "note_id",
            "required": true,
            "schema": {
              "title": "Note Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete an operator note",
        "tags": [
          "Analytics & Export"
        ]
      }
    },
...
  }
}